Safari-611.1.10/LayoutTests/platform/ios-wk2/imported/w3c/web-platform-tests/css/css-flexbox/radiobutton-min-size-expected.txt

You should see two identical-looking lines, both with a radio button at the beginning.


Text
Text

FAIL two radio button sizes are identical assert_equals: width should be equal expected 2 but got 16
